change serviceid and node to uint64

+++ topology/slave.go
@@ -40,13 +40,13 @@ func (s *slave) OnSocketMSG(dst core.ServiceID, data ...interface{}) {
 		array := sdata.([]interface{})
 		scmd := array[0].(string)
 		if scmd == "registerNodeRet" {
-			nodeId := array[1].(uint)
+			nodeId := array[1].(uint64)
 			core.RegisterNodeRet(nodeId)
 		} else if scmd == "distibute" {
 			data := array[1].([]interface{})
 			core.DistributeMSG(s.Id, data...)
 		} else if scmd == "getIdByNameRet" {
-			id := array[1].(uint)
+			id := array[1].(uint64)
 			ok := array[2].(bool)
 			name := array[3].(string)
 			rid := array[4].(uint)
